Extract animated radio and option row into shared components
